feat(AvailableDisplay.tsx): add AvailableDisplay component
The AvailableDisplay component is added to the project. This component is responsible for displaying the available quantity of an item and allowing the user to add it to the cart. It also handles the logic for updating the quantity and showing a dialog when an item is added to the cart.