feat(ui-dsc): move OutlinedDropdown component
The OutlinedDropdown component is moved from the /src/common folder to the /src folder. This component provides an outlined dropdown box that displays a list of given values. It accepts properties such as handleChange (a function called when selecting a dropdown option), dropdownValues (an array of values to display in the dropdown), inputValue (the initial value for the dropdown), and label (the input label). The component uses Material-UI components such as FormControl, InputLabel, OutlinedInput, MenuItem, and Select to create the dropdown functionality.