feat(ui-dsc): add StrainDetailsCardHeader component
The StrainDetailsCardHeader component is added to the catalog directory in the ui-dsc package. This component is responsible for displaying the header at the top of the card on the strain details page. It includes tabs for switching between strain details and phenotypes, as well as displaying the availability of the strain and the number of phenotypes.