feat(ui-dsc): add TabPanel component to catalog
The TabPanel component is a simple wrapper for content inside of tab panels. It takes in the following properties: children (any content inside the tab panel), value (tab value), and index (index of the tab selected). The component is used to display the content of a specific tab based on the selected index.