feat(ui-dsc): add getDepositorName utility function
The getDepositorName utility function is added to the ui-dsc package. This function takes a Depositor object as input and returns a string representing the depositor's name. It concatenates the first name and last name if both are provided, otherwise it returns an empty string. This utility function can be used to display the depositor's name in the user interface.