chore(stock-center): add "type" field to package.json to specify module type feat(stock-center): add "serve" script to package.json to run the server.js file
The "type" field is added to the package.json file with the value "module" to indicate that the project is using ECMAScript modules. This allows the project to use modern JavaScript features. The "serve" script is added to the package.json file to run the server.js file, which is responsible for serving the application.