refactor(CatalogTableDisplay.tsx): remove unused lastIndex variable and update rowFunction to use strains.length instead
The lastIndex variable was unnecessary, so it was removed. The rowFunction was updated to use strains.length instead of the lastIndex variable to determine if it is the last item. Additionally, the function name abbreviateString changed to a curried function called abbreviateStringToLength so that it could be used in a pipe.