feat(strains/[id].tsx): add new file for displaying strain details
A new file, [id].tsx, has been added to the strains directory. This file imports the StrainDetailsContainer component from the @dictybase/ui-dsc package and exports it as the default component for displaying strain details. This addition allows for the implementation of a page that shows detailed information about a specific strain.