feat(AddToCartControl.tsx): add AddToCartControl component
The AddToCartControl component is added to the project. It provides a button that allows users to add items to the cart. The component uses the SecondaryButton component from the @dictybase/ui-dsc library. It also utilizes state atoms from the project's state management system (jotai) to handle adding items to the cart and tracking the current cart quantity.