feat(UnavailableButton.tsx): add UnavailableButton component to display a disabled button with a tooltip when an item is not in stock
The UnavailableButton component is added to the catalog directory. This component is used to display a disabled button with a tooltip when an item is not in stock. It uses Material-UI's IconButton and Tooltip components, as well as FontAwesomeIcon for the cart and slash icons. The button is disabled and has a default color. The size of the icon and the title for the tooltip can be customized through props.