feat(navbar): add setup file for navbar package to configure testing environment
The setup.ts file is added to the navbar package to configure the testing environment. It imports necessary dependencies for testing, extends the expect method with methods from react-testing-library, and sets up a cleanup function to be executed after each test case. This setup file ensures that the testing environment is properly configured for testing the navbar package.