feat(Brand.tsx): add Brand component to navbar package
The Brand component is added to the navbar package. It is a functional component that creates a "brand" item in the navbar. It takes in the title and href as props to create a link with the specified title and URL. The component is styled using Material-UI's styling API and supports theming.