feat(navbarItems.ts): add navbarItems module to handle navbar data and formatting functions
The `navbarItems.ts` file is a new addition to the project. It introduces a module that handles the data structure and formatting functions for the navbar items. The module exports the `navbarItems` object, which contains the data for the navbar items, as well as the `formatNavbarData` function, which converts the received navbar JSON data into the desired format for the navbar component. This module will be used to populate the navbar with the appropriate data and structure.