feat(ui-dsc): add CatalogHeader component to display the title of the catalog page
The CatalogHeader component is added to the ui-dsc package. This component is responsible for displaying the title of the catalog page. It receives a "title" prop which can be either "Strain Catalog" or "Plasmid Catalog". The component uses Material-UI's Typography component to render the title in an h1 heading style. The component also applies some custom styling using makeStyles hook from Material-UI to set the color and margin of the notice and title elements.