feat(mocks): add mockStrainQuery handler and import availableStrain from ui-dsc
The mockStrainQuery handler is added to the list of handlers in the mocks/handlers.ts file. This handler allows for mocking the response of a GraphQL query for a single strain. The availableStrain object is imported from the ui-dsc package to be used as the mocked strain data.
In the ui-dsc package, the export of mockStrain is added to the index.ts file. This allows other modules to import and use the mockStrain function for testing or mocking purposes.