refactor(AddToCartButton.tsx): separate AddToCartDialog
The AddToCartDialog component has been separated from the AddToCartButton to avoid a bug where the dialog does not appear when a users adds a strain to the cart. When a strain is added to the cart, the AddToCartButton is unmounted and the RemoveFromCart button is rendered in its place. Consequently, the AddToCartDialog did not appear since it was a child of the AddToCartButton.
The unused variables and props (setHover, setCheckedItems) have been removed from the AddToCartButton component. The setShowDialog prop has been updated to use the setShowDialog prop passed from the parent component. This improves code readability and removes unnecessary code.