feat(ui-dsc): add LinkList component to generate a list of links
The LinkList component is added to the ui-dsc package. It generates a list of links based on the passed-in array of LinkItem objects. The component takes two properties: list, which is an array of LinkItem objects, and bgColor, which specifies the color of the panel background (either gray or blue). The component renders a div with a panel class, and if the list contains an item with the name "Phenotype Ontology", it adds a Typography component with the text "Download / View" as a header. The list of links is rendered as an unordered list with each link item rendered as a list item. The type of link is determined based on the properties of the LinkItem object. If routerAware is true, the link is rendered as a Link component from react-router-dom. If newWindow is true, the link is rendered as an anchor tag with the target attribute set to "_blank". Otherwise, the link is rendered as a regular anchor tag.