feat(ui-dsc): add linkLists module to provide arrays of links for different sections
The linkLists module is added to provide arrays of links for different sections of the application. This allows for easy management and customization of the links displayed in the UI. The module exports four arrays: downloadLinks, infoLinks, materialsLinks, and miscLinks, each containing objects with properties such as name, to, routerAware, and newWindow. These arrays can be used to populate the appropriate sections of the UI with the corresponding links.