feat(index.tsx): add homepage component for Dicty Stock Center (DSC)
The index.tsx file in the pages directory is added to introduce the homepage component for the Dicty Stock Center (DSC) application. This component includes various UI elements such as Grid, Typography, Helmet, and LinkList from the @dictybase/ui-dsc package. It also imports link lists for download, information, materials, and miscellaneous links. The component is wrapped in a Helmet component to set the page title and meta description. The component is exported as the default export.