feat(AvailableDisplay.tsx): add functionality to remove item from cart and update UI accordingly
The unused import of the OutlinedDropdown component has been removed to clean up the code. The AvailableDisplay component now has the ability to remove an item from the cart by clicking the "Remove From Cart" button. The UI is updated accordingly to show the "Remove From Cart" button when an item is already in the cart, and the "Add to Cart" button when the item is not in the cart.