chore(dicty-graphql-schema): update eslint rules in package.json
The "@graphql-eslint/naming-convention" rule is turned off to allow for more flexibility in naming conventions. The "@graphql-eslint/require-description" rule is also turned off to avoid enforcing descriptions for all GraphQL types and fields. Additionally, the "@graphql-eslint/no-typename-prefix" rule is turned off to allow for the use of typename prefixes. The "@graphql-eslint/strict-id-in-types" rule is updated to include additional accepted ID names and types. Finally, the list of accepted types in the "@graphql-eslint/enum-value-should-be-camel-case" rule is updated to include "DeleteContent", "DeletePermission", "DeleteRole", "DeleteStock", and "DeleteUser".