feat(node-ci-dev.yml): add call-docker-build job to trigger a custom docker build and push workflow
The workflow name is updated to "Continuous integration" to better
reflect its purpose. The runs-on value is changed to "ubuntu-latest" to
ensure compatibility with the latest version of Ubuntu. The node-version
is updated to "lts" to ensure better stability and support. The
codecov-action version is updated to v3 to take advantage of the latest
features provided by codecov. Additionally, a new job called
"call-docker-build" is added to trigger a custom docker build and push
workflow.