chore(ci.yml): add continuous integration workflow for unit tests and Docker build
A new continuous integration workflow has been added to the repository.
This workflow is triggered on push events to the develop branch. It
consists of two jobs: "test" and "call-docker-build". The "test" job
runs unit tests using Node.js 18 on an Ubuntu environment. It sets up
Node.js, checks out the code, renames a config file, installs
dependencies, runs unit tests, and uploads the test coverage to Codecov.
The "call-docker-build" job is dependent on the "test" job and uses a
separate workflow from the dictyBase repository to build and push a
Docker image for the stock-center application. The workflow uses secrets
from the repository and provides the necessary inputs for the Docker
build. This workflow aims to ensure the quality of the codebase through
automated testing and to streamline the Docker image build process.