feat(ui-dsc): add convertToPayerField utility function
The `convertToPayerField` utility function is added to the `utils` directory in the `ui-dsc` package. This function takes a `name` parameter and converts it to a payer field by capitalizing the first character and appending "payer" to it. This utility function can be used to transform field names to match the naming convention used for payer fields in the application.