test(ui-dsc): update PaymentMethodRadioGroup test to remove unnecessary assertions
The unnecessary assertions for the "Purchase Order (PO)" payment method have been removed because the default value of the radio group is set to Purchase Order, so clicking on the on the purchase order radio button initially will not fire the onChange handler. The number of times the mockSetPaymentMethod function is called has been adjusted accordingly.