feat(PaymentMethodRadioGroup.tsx): add logic to reset and set value of purchaseOrderNum field based on selected payment method
The PaymentMethodRadioGroup component now includes logic to reset and set the value of the purchaseOrderNum field based on the selected payment method. If the payment method is set to "PURCHASE_ORDER_NUMBER", the purchaseOrderNum field is reset. Otherwise, the purchaseOrderNum field is set to "N/A" with the option to mark it as dirty. This logic ensures that the purchaseOrderNum field behaves correctly based on the selected payment method.